+++ Server/Models/Monsters/JinamStoneGate.cs
@@ -97,7 +97,7 @@ namespace Server.Models.Monsters
         {
         }
 
-        public override int Attacked(MapObject attacker, long power, Element element, bool canReflect = true, bool ignoreShield = false, bool canCrit = true, bool canStruck = true)
+        public override long Attacked(MapObject attacker, long power, Element element, bool canReflect = true, bool ignoreShield = false, bool canCrit = true, bool canStruck = true)
         {
             return 0;
         }
