+++ Server/Models/Monsters/SunFeralWarrior.cs
@@ -13,9 +13,9 @@ namespace Server.Models.Monsters
         public bool CanSpawn = true;
 
 
-        public override int Attacked(MapObject attacker, long power, Element element, bool canReflect = true, bool ignoreShield = false, bool canCrit = true, bool canStruck = true)
+        public override long Attacked(MapObject attacker, long power, Element element, bool canReflect = true, bool ignoreShield = false, bool canCrit = true, bool canStruck = true)
         {
-            int result = base.Attacked(attacker, power, element, canReflect, ignoreShield, canCrit);
+            long result = base.Attacked(attacker, power, element, canReflect, ignoreShield, canCrit);
 
             if (!CanSpawn || result <= 0) return result;
